Vision system refactor
This changes the recursive shadowcast implementation to scan octants
from the cardinal to ordinal axis, rather than always scanning left to
right, top to bottom. This allows octants to be encapsulated as their
own class rather than a collection of arguments to recursive shadowcast.
It also allows for an arguably nicer implementation of cell sides for
proper lighting support. Additionally, it will be easier to support
viewing a range of angles rather than always seeing 360 degrees.